# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BoscovFilmes API",
    "version": "1.0.0",
    "description": "Documentação da API do BoscovFilmes"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor local"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "summary": "Login de usuário",
        "description": "Realiza o login e retorna um token JWT.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "usuario@exemplo.com" },
                  "senha": { "type": "string", "example": "123456" }
                },
                "required": ["email", "senha"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        }
      }
    },
    "/me": {
      "get": {
        "summary": "Obter dados do usuário logado",
        "description": "Retorna os dados do usuário autenticado.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Dados do usuário retornados com sucesso"
          },
          "401": {
            "description": "Token não fornecido"
          },
          "403": {
            "description": "Token inválido"
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "Cadastro de usuário",
        "description": "Cria um novo usuário.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": { "type": "string", "example": "João da Silva" },
                  "email": { "type": "string", "example": "novo@exemplo.com" },
                  "senha": { "type": "string", "example": "123456" },
                  "dataNascimento": { "type": "string", "format": "date", "example": "2000-01-01" }
                },
                "required": ["nome", "email", "senha", "dataNascimento"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "409": {
            "description": "E-mail já cadastrado"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
} 